Extract buffet into its own crate
