+++ crates/hring-buffet/Cargo.toml
@@ -1,0 +1,31 @@
+[package]
+name = "hring-buffet"
+version = "0.1.0"
+edition = "2021"
+license = "MIT OR Apache-2.0"
+repository = "https://github.com/hapsoc/hring"
+documentation = "https://docs.rs/hring-buffet"
+readme = "README.md"
+description = """
+Buffer management for the `hring` crate.
+"""
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[features]
+miri = []
+
+[dependencies]
+eyre = "0.6.8"
+http = "0.2.8"
+memchr = "2.5.0"
+memmap2 = { version = "0.5.8", default-features = false }
+nom = "7.1.3"
+pretty-hex = "0.3.0"
+thiserror = { version = "1.0.38", default-features = false }
+tokio = { version = "1.25.0", features = ["sync", "macros"] }
+tokio-uring = "0.4.0"
+tracing = "0.1.37"
+
+[dev-dependencies]
+pretty_assertions = "1.3.0"
