+++ crates/hring-buffet/src/io.rs
