+++ crates/hring-buffet/src/piece.rs
