+++ crates/hring/Cargo.toml
@@ -15,10 +15,10 @@ byteorder = "1.4.3"
 enum-repr = "0.2.6"
 enumflags2 = "0.7.5"
 eyre = { version = "0.6.8", default-features = false }
+hring-buffet = { version = "0.1.0", path = "../hring-buffet" }
 hring-hpack = { path = "../hring-hpack" }
 http = "0.2.8"
 memchr = "2.5.0"
-memmap2 = { version = "0.5.8", default-features = false }
 nom = { version = "7.1.3", default-features = false }
 pretty-hex = { version = "0.3.0", default-features = false }
 smallvec = { version = "1.10.0", default-features = false, features = ["const_generics", "const_new", "union"] }
