+++ crates/hring/src/types/method.rs
@@ -1,6 +1,6 @@
 use std::fmt;
 
-use crate::{Piece, PieceStr};
+use hring_buffet::{Piece, PieceStr};
 
 /// An HTTP method, see https://httpwg.org/specs/rfc9110.html#methods
 #[derive(Clone)]
